requirements.txt
# Core dependencies
pyspark==3.5.1
pymongo==4.6.0
newsapi-python==0.2.7
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0

# NLP and sentiment analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard
streamlit==1.29.0

# Additional dependencies
numpy==1.26.2
scipy==1.11.4
